refactor(script): Phase 4 Step 4 - Extract artifact_management.py
- Created src/script/artifact_management.py (101 lines)
- Extracted move_script_files_to_artifacts() function
- Manages browser script artifact collection and metadata
- Added import to script_manager.py for backward compatibility
- script_manager.py: 1,646 → 1,563 lines (-83 lines, 5.0%)
- Total with new module: 1,664 lines (+18 lines from structure)
